+++ lib/Cascade_Graphics/src/Vulkan_Wrapper/descriptor_set_manager.cpp
@@ -54,7 +54,7 @@ namespace Cascade_Graphics
             descriptor_set_layout_create_info.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_SET_LAYOUT_CREATE_INFO;
             descriptor_set_layout_create_info.pNext = nullptr;
             descriptor_set_layout_create_info.flags = 0;
-            descriptor_set_layout_create_info.bindingCount = descriptor_set_layout_bindings.size();
+            descriptor_set_layout_create_info.bindingCount = static_cast<uint32_t>(descriptor_set_layout_bindings.size());
             descriptor_set_layout_create_info.pBindings = descriptor_set_layout_bindings.data();
 
             VALIDATE_VKRESULT(vkCreateDescriptorSetLayout(*m_logical_device_ptr->Get_Device(), &descriptor_set_layout_create_info, nullptr, &descriptor_set_data_ptr->descriptor_set_layout), "Vulkan: Failed to create descriptor set layout");
@@ -83,7 +83,7 @@ namespace Cascade_Graphics
             descriptor_pool_create_info.pNext = nullptr;
             descriptor_pool_create_info.flags = 0;
             descriptor_pool_create_info.maxSets = 1;
-            descriptor_pool_create_info.poolSizeCount = descriptor_pool_sizes.size();
+            descriptor_pool_create_info.poolSizeCount = static_cast<uint32_t>(descriptor_pool_sizes.size());
             descriptor_pool_create_info.pPoolSizes = descriptor_pool_sizes.data();
 
             VALIDATE_VKRESULT(vkCreateDescriptorPool(*m_logical_device_ptr->Get_Device(), &descriptor_pool_create_info, nullptr, &descriptor_set_data_ptr->descriptor_pool), "Vulkan: Failed to create descriptor pool");
@@ -181,7 +181,7 @@ namespace Cascade_Graphics
                 }
             }
 
-            vkUpdateDescriptorSets(*m_logical_device_ptr->Get_Device(), descriptor_set_data_ptr->write_descriptor_sets.size(), descriptor_set_data_ptr->write_descriptor_sets.data(), 0, nullptr);
+            vkUpdateDescriptorSets(*m_logical_device_ptr->Get_Device(), static_cast<uint32_t>(descriptor_set_data_ptr->write_descriptor_sets.size()), descriptor_set_data_ptr->write_descriptor_sets.data(), 0, nullptr);
         }
 
         void Descriptor_Set_Manager::Create_Descriptor_Set(std::string resource_grouping_label)
